analysis_chat/pyproject.toml
[project]
name = "analysis-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "cardrapier",email = "cardrapier@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka (>=2.8.0,<3.0.0)",
    "nats-py (>=2.9.0,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

